[CI/CD] Refactor Cypress workflow
